#42 - Update Logging System
The 'debug' library had some issues that were causing tests to fail in interesting ways. When we ran tests with DEBUG=app*,gentest; the expected result would be that the tests stayed stable, but we would see application debug statements. The debug library writes content to stderr by default, and the global override for using stdout is buggy. This caused the generator tests to fail, because they expect stderr to be empty. Because of this, `debug` is not meeting our needs, and we're switching the generated code to using log4js.  Generator code will continue to use debug.

Also, removing the startupHoopManager test, which is resulting in some bizarre issues in the sequelize generator.